fix OffsetIndex FirstRowIndex

+++ writer/writer.go
@@ -443,7 +443,7 @@ func (pw *ParquetWriter) Flush(flag bool) error {
 					offsetIndex.PageLocations = append(offsetIndex.PageLocations, pageLocation)
 
 					//firstRowIndex += int64(page.Header.DataPageHeader.NumValues)
-                    firstRowIndex += page.NumRows
+                    firstRowIndex += int64(page.NumRows)
 				}
 
 				data := rowGroup.Chunks[k].Pages[l].RawData
